Extracted watch command to separate file
